layout: add dock_popup layout
This layout is meant to be shown when the user clicks a marker. It's
implemented with a TableLayout that currently shows the name of the
station, the number of bikes and docks, and a temporary button to close
the popup.

Longterm, this button might be replaced with buttons to add and remove
stations from the watchlist