activity: map: refactor: add a `BixiStation` directly as a Marker tag
I believe this will help in the longterm since we have two issues, one
is to be able to access the stations easily via ID, since that's what
we get from the API, and that will be used when updating stations.

On the other hand, we have the issue of keeping track of the Markers
since it will be nice to have them visually change depending on the
amount of available bikes. The problem with the current approach is that
it actually takes some time for the map to update the Markers. So this
adds the requirement of having to unblock the UI thread enough so that
the map continues loading. To do this we actually need to order the
Markers (by distance from the user), so having the BixiStation directly
accessible from the Marker simplifies things.

All this is working under the assumption that classes are passed by
reference in Kotlin, which means that modifying BixiStations from either
the Marker tag or from their MutableMap will have the same effect