bixi api: make the `BixiStation` class sortable
The idea behind this is to be able to tell which are the stations that
are near the user. To achieve this, a "static" variable had to be added
so that the location of the user can be known by all instances.

The location of the user is only updated on load.

In the future it could also be updated whenever we get new data from
the API, so that updates in color (or anything else) in the markers can
be done in the same way starting by stations closer to the user and
ending with those further away.