bixi api: add class variable to keep track of current hue
The hue is defined as a value between 0 (Red) and 120 (Green), where 0
means the lack of available bikes and green means that the station is
full of bikes.

Since the class already has access to this information, it makes sense
to have it keep track of its current hue even though that's only a user
facing property. This is useful because when we update the station
information after an API call, we don't want to automatically update
the hue.

The idea is to only update Markers on need, so comparing the current hue
with the new hue would save some time on the UI thread and make the app
a bit more responsive