activity: map: add colors on the markers depending on bike availability
This commit simply implements the user-facing side of the new colors for
the markers.

This happens in a separate block because we want first to be sure that
all markers have been placed. They have already been ordered, so we can
simply iterate through them with a small `sleep` to unblock the UI and
keep responsiveness.

In practice, at the current setting of 10ms sleep, there is still a
small hiccup at the beginning, but responsiveness comes back pretty
quickly. The longterm plan is to make this a configurable option, so
that people can choose to have this load at say 100ms/Marker, because
the Markers will still be created in their default red color quickly,
and the number of available bikes can still be checked before the colors
finish being updated