bixi station: reduce the color range
This commit reduces the color range to be shown in the markers from
0-120 to 30-90. The idea is that stations that are actually full or
empty should be a bit more visible. This still gives some nice range for
mid values.

Maybe something similar could be done at some point for "almost full" or
"almost empty" stations, but not in the plans for now