activity: map: add location updates and callback
This commit starts location updates on app start (or when location
permission is granted). For now it's set to request every 10s, with high
accuracy.

In practice, by looking at the emulator, it seems to be faster, but in
practice there is no difference. Will need to check at some point how
this affects battery usage, etc. But we're far from that

For now, the callback itself only updates the static variable of the
BixiStation class