logic: identify when user is close to a station
This is simply done by checking the distance between the user and the
closest station. For now, there is a hardcoded check for 3m or less, but
this might change depending on tests.

On top of that, to avoid sorting the bixi stations each time there is a
location update, first we check if the current location is further than
100m. Also hardcoded for now, and also could be modified after testing.

Finally, we check only the 10 closest (after the last sort) stations.
This is because since we're not re-sorting every time, it is possible
that it's not always the first station the one that is the closest.

All these hardcoded values were roughly decided by checking on a map
and seeing rough distances bewteen stations even in crowded areas