lint: move object implementation to its definition
This is simply to organize the code a bit more. Since the implementation
of the LocationCallback object is removed enough from the place it was
before. It makes sense to have it a the top-level of the activity class.

Besides this, this commit also adds a few comments for clarity, and also
adds a `fastestInterval` of 5s in the location requests. This is simply
to avoid updating too often since we don't really need that