logic: change storage of user tracked stations to a list of BixiStation
Previously we only kept the ids, but this had other issues since at some
point we want to be able to compare the stored values to the incoming
values after an API call.

The idea here is that we keep copies of the stations in the `userDocks`
list, and we will compare against this when updating the main station
map. In this way, we can update the main map after each API call without
having to worry about losing the ability to compare station statuses.