logic: tracking: add tracking skeleton code
This commit adds the skeleton of what will be the tracking code. We are
simply starting a CountDownTimer that will tick every 30s. The idea is
to handle the updates from the station here, and notify the user if
needed.

For now there is no way to stop the tracking. What will probably be
implemented is to have a notification with a stop button there.

This commit also removes the closestStations variable since we are
simply using the sorted stations instead.