location: remove everything related to user location
Turns out there is an "Activity Recognition API", which can be used to
identify when the user starts biking.

This means that the location approach is not needed anymore, and it can
be removed completely. That approach also wasn't ideal since we would
need constant location updates to know when the user was close to a
station, and in practice this wasn't feasible, since background services
handled by `WorkerManager` have a minimum period of 15 minutes, which
way too much for what we're trying to do. The other option was to use
a proper foreground service, but that includes the need for a constant
notification which is annoying, and would probably drain the battery to
call for location updates that often.

The ability to sort the docks and a fixed "user location" was left since
that's nice when loading up the application to have some sort of order
in which the stations update themselves. And also, since the map is only
used to have a place for the user to pick their stations, responsiveness
becomes a bit more important.