logic: ensure that user docks are loaded when starting tracking
This is needed since we don't know at which point the tracking will be
starting. It could (will) happen after the user has closed the activity
and removed it from the active applications list. This means that the
applications ressources have been mostly freed and we're not guaranteed
to have data populated in the list in RAM. So to ensure we always have
valid data, we fetch and populate the list again when starting tracking.

Minor formatting change also included in this commit