logic: add text to speech when notifying the user
This commit adds a basic voiced message to go with the notification.
Since the idea is that the user will be biking, it makes sense to use
TTS instad of relying on the notifications.

Note that some quick testing revealed that music does get automatically
lowered when TTS is active, so there's nothing additional to do on that
side.