utils: add helper functions for station status and TTS
This commit adds a couple of helper functions.

One is to simply know when there has been a change in the status of any
of the tracking station.

This is done by checking against a list of currently unavailable IDs.
Because of this, we won't need anymore to keep a copy of the user-
tracked stations, and we can simply keep a direct link to the original
ones in the BixiApiHandler