logic: update handling of user notifications
This commit implements the changes from the past few commits. The idea
now is that we keep track only of the unavailable IDs, and notify the
user accordingly once they change.

Since in practice, the notifications will not be looked at, we also keep
track of the changes only compared to the previous iteration.

We notify the user via text to speech whenever any of their tracked
stations either becomes full, or stops being full.