utils: favorites: add initialization on button press
This handles all the initialization logic for the favorites view. This
is all done in the callback for the favorites button click.

Note that we're using a popout instead of a proper activity. The
reason for this is that we don't want to reload the map every time,
since it's very time consuming especially for the markers. So a popout
actually means that the map continues loading in the background, and
the transitions are seamless between the two views.