layout: add background fragment
This layout is needed since we're not going to be using a proper
activity to display the settings page.

Instead, we are relaying on applying a couple of fragments on top of
the map. The "base" fragment is this layout, and the "top" fragment
is the one genreated by the settings xml.

This is a temporary (I hope) approach. Since the proper approach would
be to use Activities, and use the Bundle data to avoid reloading the
map every time. That approach still needs some investigation, but it
should be the cleanest approach.