configuration: change base configuration folder and add getters
This commit basically unifies the default configuration folder used
by the Preference class.

Should be noted that for now the folder is hardcoded, so this could
potentially cause issues if the default folder changes.

On top of this, this commit simply adds ease of use functions for the
implemented settings.