api: update to the General Bikeshare Feed Specification (gbfs)
This update requires some moderate refactoring on the api and the
ShareStation class.

This is mostly because the gbfs now provides the information needed in
two separate endpoints.

One is only for station static information. This means usually things
like location, total capacity, name, etc.

And the other one is for station "dynamic" status, which includes the
availability of bikes and docks.

How this affects the previous logic is that now we can't create the
`ShareStation`s in one go, so we need an "empty" constructor, and then
set all the parameters as we go along.

This also means separating the "load docks" function into "load info"
and "load status", which luckily wasn't too bad since we already had
similar logic for "load" and "update".

The one caveat is that the "status" function assumes that the station
map has already been populated, and ignores any received IDs that
weren't previously mapped. This shouldn't be an issue in practice, since
we load the info on app startup, but also again on tracking start. So
in practice it will always have valid data before going into the
tracking loop.