logic: attempt to reload docks if map is empty
This is a safety measure in case of when starting the tracking, a
network error occurred and the stations couldn't be loaded. So, at
every tick of the timer we ensure there is at least some data in the
map, and if there is none, we reload the docks and user stations.