city utils: move city information to a separate object
As cities are added, the utils file would have become too big, so this
helps keeping things separate.

Also, we're using a proper class instead of a simple String, since we
need to keep track of additional information such as the default
location.

On top of this, we make it a map of Int -> CityInfo since the integer
will be used as an ID to compared with the one set in the settings page.