api: use the base gbfs url to get the station info ones
After starting to work on adding the cities listed in the gbfs repo, I
realized that not all of them adhere to the url format I expected. More
specifically, not all of them had a properly formatted url with the
.json extension at the end. There were a few other inconsistencies like
the base gbfs and the information urls were not in the same base host.

The easiest way to solve this is to actually follow the spec the way it
was meant to be used. So this commit changes a few things to do that.

Now we only keep track of the base gbfs.json url. Because of this, since
no more string concatenations are needed, we can make it a proper `URL`
on the CityInfo class.

We keep the info url in ram, which means we also need to ensure they are
loaded properly. So before every call to the `load` stations, we also
load the urls.

This also opens the possibility to easily add other of the optional urls
in the gbfs spec, but none are in the plans for now.