city utils: add support for cities in listed in the gbfs systems.csv
This commit adds support for most of the cities that were listed in the
master systems.csv database in the gbfs repository [1].

The requirements for a city to be added were as follows:
- Compliant to the gbfs and listed in the systems.csv.
  + Some cities that weren't strictly compliant were added since it did
    not require significant changes.
- Have a dock-based bike share system.
  + Free-use systems are not within the scope of this app.
- Have more than 2 docks which aren't close together.
  + No need to add cities that have docks in a single location meant
    mostly for touristing.
- Have an appropriately ssl verified hostname.
  + Notably, some cities in Great Britain had to be dropped since
    attempting to retrieve data from their system threw a SSL exception
    even with plain text allowed for those domains.

Each city entered has the following information:
- 2-letter country code
- Name of the city in (mostly) latin alphabet
- Rough coordinates for where most of the stations are placed
- The base gbfs url

Note that after this commit, every city entered is forever tied to its
id. Future additions will not break this, and it will make it so that
the arrays.xml values is not ordered anymore since we need to keep
alphabetical order.

There was a significant amount of effort putting all this together and
validating properly for each city.

Special thanks to Shrisha "hawtsaws" M. and Ingrid "nope" R. for their
help sorting this out.

[1]: https://github.com/NABSA/gbfs/commit/37dab55460469bf0bdb7ee4d4e10a4d879737b2d