activity: map: increase visibility of the settings button
It seems that adding a button to a fragment makes it so that the button
is somewhat translucent. One solution to this is to have a solid
background, like it's done in the favorites button.

However, the solid background would go beyond the borders of the icon,
so what we do instead is to set both the foreground and the background
to the same image, which allows for some "solid" background to the icon.

Also, some color overlay is added so that it goes with the rest of the
app.