logic: ensure that variables are properly loaded
Previously we skipped the re-loading of the the urls if the map had
data. However, my guess is that android would free memory by stages,
so it is possible that the map data was kept while the URLs weren't.

There's also other changes including the order in which variables
are loaded, and some things that weren't initialized at all when
starting from the activity recognition.