logic: skip tracking tick if network error or system shutdown
To identify a network error, we simply check no stations were loaded. In
that case, we assume there was a network error and will attempt to load
again in the next tick.

To identify a system shutdown, we check if no stations are available. It
is likely that the user isn't using a shared bike at all, but since it
could be a temporary system shutdown triggered after the start, it's
better to just try again at the next tick. For seasonal shutdowns, the
expected approach would be to remove the tracking at the settings level.