favorites: allow dragging entries to reorder
We implement the drag directions on the ItemTouchHelper for Up and Down.
And fill up the onMove function accordingly.

However, because this function is called too often, it makes sense to
override the clearView as well, which is called once the action (drag
or swipe) is over, and we save to the configuration file here.