activity: popup: add nickname handling
Allow the user to set a nickname per station. For now, a button is added
at the bottom, but the visuals will be reworked a bit.

The nickname displays instead of the normal station name, and the
dialog to edit the text gives the option to remove the nickname if
required.