activity: map: update map visuals on relevant settings change
The load of the map now happens as follows:
1. Remove all existing markers
2. Load new markers
3. Remove all inactive stations
4. Reload all inactive stations if required by the settings
5. Load all colors if required by the settings

The reason for step 3 and 4 is to simplify the logic, since the visual
load function repeats step 3 through 5 without having to keep track
specifically of the inactive stations.

Basically, step 4 is to ensure a consistent state before performing
the visual changes as required by the settings.