map: use custom icons
The main advantage of using custom icons is that the load happens
significantly faster. In addition to that the following nice changes
were done:

- "Fill" the icon instead of using a color code
- Change the color of the border when the station is completely full
  or empty
- Can actually show grey icons for unavailable stations

The border and shape of the icon are basically a vector path, which is
generated programmatically, since inflating an xml was unnecessary and
it needed to happen for every single station. This path is also used
to clip the canvas, and with the shape at the canvas level, it can be
filled up with basic rectangles.