city utils: generate sources from csv (and drop support for many cities)
"many" here really means all except Montreal. Support for other cities
would be on a per-need basis, or if this application ever becomes
popular enough.

The sources that are now being generated automatically are CityUtils
and the resource array. This simplifies eventually adding more cities,
since we only need to modify the csv.